requirements.txt
sqlalchemy
psycopg2-binary
grpcio
grpcio-tools
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-grpc
pylint
flake8
grpcio-reflection
fastapi
uvicorn
jinja2
python-multipart